Move error handling tests to a separate file
